Separate out common code between view versions into utils
