+++ src/polyswarmd/views/v1/staking.py
@@ -3,10 +3,11 @@ import logging
 import fastjsonschema
 from flask import Blueprint, g, request
 
+from polyswarmd.utils import eth
 from polyswarmd.utils.decorators.chains import chain
+from polyswarmd.utils.eth import build_transaction
 from polyswarmd.utils.response import failure, success
-from polyswarmd.views.v1 import eth
-from polyswarmd.views.v1.eth import build_transaction
+
 
 logger = logging.getLogger(__name__)
 staking: Blueprint = Blueprint('staking', __name__)
