+++ src/polyswarmd/websockets/messages.py
@@ -135,7 +135,7 @@ class MetadataHandler:
     def initialize(cls):
         """Create & assign a new implementation of _substitute_metadata"""
         from polyswarmd.app import app
-        from polyswarmd.views.v1.bounties import substitute_metadata
+        from polyswarmd.utils.bounties import substitute_metadata
         config: Optional[Dict[str, Any]] = app.config
         ai = config['POLYSWARMD'].artifact.client
         session = FuturesSession(adapter_kwargs={'max_retries': 3})
